+++ doc/version_history.rst
@@ -6,6 +6,17 @@
 Version History
 ###############
 
+v1.5.0
+======
+
+Prepare for salobj 7.
+
+Requires:
+
+* ts_salobj 7.0
+* ts_idl 3.1
+* IDL file for WeatherStation from ts_xml 8.0
+
 v1.4.1
 ======